Added Refinery Manager class and sample program
Refinery manager has LoadRefineries function to transfer ores to the
refinery in small stacks so that one ore doesn't clog up the system.
First argument is a group name of cargos where ores are stored
Second argument is the group name of refineries.
Third argument is the base stack size, these are multiplied by the
numbers defined in RefineryManagerSettings and can be customised.