Inventory: New Features
Added overloads for CountItems so you can now pass list of IMyInventorys
or just a single IMyInventory (in addition to list of IMyTerminalBlock).
Added MoveItemAmount to enable moving of a certain number of number of
items. Fails silently if for any reason can't move items.
Added Additional overload for GetFercentFull to take single IMyInventory
as argument.
Added MaxInventoryPercent static var as an option.